refactor #101 Profile remote 폴더 제거
