+++ lib/features/profile/bindings/profile_binding.dart
@@ -1,13 +1,13 @@
 import 'package:debateseason_frontend_v1/core/network/dio_client.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/profile_data_source.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/users_logout_data_source.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/users_withdraw_data_source.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/remote/profile_repository_impl.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/remote/users_logout_repository_impl.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/remote/users_withdraw_repository_impl.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/profile_repository.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/users_logout_repository.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/users_withdraw_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/profile_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/users_logout_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/users_withdraw_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/profile_repository_impl.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/users_logout_repository_impl.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/repository_impls/users_withdraw_repository_impl.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/profile_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/users_logout_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/users_withdraw_repository.dart';
 import 'package:debateseason_frontend_v1/features/profile/presentation/view_models/profile_view_model.dart';
 import 'package:get/get.dart';
