+++ lib/features/profile/data/data_sources/community_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/profile/data/models/remote/response/community_res.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/models/response/community_res.dart';
 import 'package:debateseason_frontend_v1/utils/base/base_res.dart';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
