+++ lib/features/profile/data/data_sources/profile_data_source.dart
@@ -1,5 +1,5 @@
-import 'package:debateseason_frontend_v1/features/profile/data/models/remote/request/profile_req.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/models/remote/response/profile_res.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/models/request/profile_req.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/models/response/profile_res.dart';
 import 'package:debateseason_frontend_v1/utils/base/base_res.dart';
 import 'package:debateseason_frontend_v1/utils/base/nullable_base_res.dart';
 import 'package:dio/dio.dart';
