+++ lib/features/profile/data/data_sources/users_logout_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/profile/data/models/remote/request/users_logout_req.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/models/request/users_logout_req.dart';
 import 'package:debateseason_frontend_v1/utils/base/nullable_base_res.dart';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
