+++ lib/features/profile/data/mapper/community_mapper.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/profile/data/models/remote/response/community_res.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/models/response/community_res.dart';
 import 'package:debateseason_frontend_v1/features/profile/domain/entities/community_entity.dart';
 
 class CommunityMapper {
