+++ lib/features/profile/data/repository_impls/profile_nickname_check_repository_impl.dart
@@ -1,5 +1,5 @@
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/profile_nickname_check_data_source.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/profile_nickname_check_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/profile_nickname_check_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/profile_nickname_check_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/nullable_base_res.dart';
 
 class ProfileNicknameCheckRepositoryImpl
