+++ lib/features/profile/data/repository_impls/profile_repository_impl.dart
@@ -1,8 +1,8 @@
 import 'package:debateseason_frontend_v1/core/services/shared_preferences_service.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/profile_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/profile_data_source.dart';
 import 'package:debateseason_frontend_v1/features/profile/data/mapper/profile_mapper.dart';
 import 'package:debateseason_frontend_v1/features/profile/domain/entities/profile_entity.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/profile_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/profile_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 
 class ProfileRepositoryImpl implements ProfileRepository {
