+++ lib/features/profile/data/repository_impls/users_withdraw_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:debateseason_frontend_v1/core/services/secure_storage_service.dart';
 import 'package:debateseason_frontend_v1/core/services/shared_preferences_service.dart';
-import 'package:debateseason_frontend_v1/features/profile/data/data_sources/remote/users_withdraw_data_source.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/users_withdraw_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/data/data_sources/users_withdraw_data_source.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/users_withdraw_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 
 class UsersWithdrawRepositoryImpl implements UsersWithdrawRepository {
