+++ lib/features/profile/presentation/view_models/profile_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:debateseason_frontend_v1/features/profile/domain/entities/profile_entity.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/profile_repository.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/users_logout_repository.dart';
-import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/users_withdraw_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/profile_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/users_logout_repository.dart';
+import 'package:debateseason_frontend_v1/features/profile/domain/repositories/users_withdraw_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 import 'package:debateseason_frontend_v1/utils/logger.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
